+++ NEWS.md
@@ -1,5 +1,6 @@
 # flapper 1.0.0
 
+* Update `get_mvt_mobility_from_archival()` with `step_check` argument to check for regular time series;
 * Improve `dist_btw_receivers()` to handle lon/lat and planar coordinates and return dataframes or matrices;
 * Fixes to `get_detection_clumps()` and `lcp_over_surface()` after dependencies update;
 * Set up testing with `testthat`;